Ping now uses overwritten session metadata
